fixed test generation
