+++ src/meson.build
@@ -1,7 +1,9 @@
 # src build file
 #============================================================================
 
+
 hatchet_inc += include_directories('.')
+
 
 src_dirs = [
     'libhatchet',
@@ -13,13 +15,14 @@ src_dirs = [
     'hatchet'
 ]
 
-if (get_option('with_tools'))
-    src_dirs += 'tools'
-endif
-
 
 if (get_option('with_crashreporter'))
     src_dirs += 'crashreporter'
+endif
+
+
+if (get_option('with_tools'))
+    src_dirs += 'tools'
 endif
 
 
@@ -31,8 +34,3 @@ endif
 foreach dir : src_dirs
     subdir(dir)
 endforeach
-
-
-
-
-
