+++ src/tests/TestResult/main.cpp
@@ -1,0 +1,31 @@
+/* === This file is part of Tomahawk Player - <http://tomahawk-player.org> ===
+ *
+ *   Copyright 2013, Dominik Schmidt <domme@tomahawk-player.org>
+ *
+ *   Tomahawk is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ *   Tomahawk is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with Tomahawk. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "TestResult.h"
+
+int main( int argc, char** argv)
+{
+    QCoreApplication app( argc, argv );
+
+    #define TEST( Type ) { \
+        Type o; \
+        if (int r = QTest::qExec( &o, argc, argv ) != 0) return r; }
+
+    TEST( TestResult );
+    return 0;
+}
