+++ src/tests/TestServent/main.cpp
@@ -1,26 +1,22 @@
-/* === This file is part of Hatchet Player - <http://hatchet-player.org> ===
+/* === This file is part of Tomahawk Player - <http://tomahawk-player.org> ===
  *
  *   Copyright 2013, Dominik Schmidt <domme@tomahawk-player.org>
  *
- *   Hatchet is free software: you can redistribute it and/or modify
+ *   Tomahawk is free software: you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
  *   the Free Software Foundation, either version 3 of the License, or
  *   (at your option) any later version.
  *
- *   Hatchet is distributed in the hope that it will be useful,
+ *   Tomahawk is distributed in the hope that it will be useful,
  *   but WITHOUT ANY WARRANTY; without even the implied warranty of
  *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *   GNU General Public License for more details.
  *
  *   You should have received a copy of the GNU General Public License
- *   along with Hatchet. If not, see <http://www.gnu.org/licenses/>.
+ *   along with Tomahawk. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include <QtTest>
-#include <QtCore>
-
-#include "Test@HATCHET_TEST_CLASS@.h"
-#include "moc_Test@HATCHET_TEST_CLASS@.cpp"
+#include "TestServent.h"
 
 int main( int argc, char** argv)
 {
@@ -30,6 +26,6 @@ int main( int argc, char** argv)
         Type o; \
         if (int r = QTest::qExec( &o, argc, argv ) != 0) return r; }
 
-    TEST( Test@HATCHET_TEST_CLASS@ );
+    TEST( TestServent );
     return 0;
 }
